chore(dal,sdf,web): expose configuration diagram node "category"
Primary:
- Expose "category" to configuration diagram nodes in order to easily
  discern the logo that should be present on the card
  - No longer based of the node title prefix

Builtins:
- Ensure doc link setting only occurs in the helper for schema builtins
  (optimization)
- Rename "butane" to "coreos_butane"
- Ensure "func" builtin module is private and anything required to be
  public is exposed via "pub use"
- Move main "kubernetes_deployment" builtin migration function to the
  parent module (now matches the behavior of all other builtin schema
  entrypoint functions)
  - Rename "kubernetes_deployment" module to
    "kubernetes_deployment_spec" since it now contains the spec (and all
    fields below it)

Misc:
- Add "SQUASING_YOUR_CHANGES" doc
- Move UiMenu list call into its own method from "node_menu"
- Move schema builtin test harness prop map creation logic for specific
  builtins to their own private methods
- Modified doc comments across the dal due to recent changes
