+++ lib/dal/src/builtins/schema/docker.rs
@@ -33,11 +33,11 @@ async fn docker_hub_credential(ctx: &DalContext) -> BuiltinsResult<()> {
         .set_component_kind(ctx, ComponentKind::Credential)
         .await?;
 
-    let (mut variant, root_prop) = SchemaVariant::new(ctx, *schema.id(), "v0").await?;
-    variant.set_color(ctx, Some(0x1e88d6)).await?;
+    let (mut schema_variant, root_prop) = SchemaVariant::new(ctx, *schema.id(), "v0").await?;
+    schema_variant.set_color(ctx, Some(0x1e88d6)).await?;
 
     schema
-        .set_default_schema_variant_id(ctx, Some(*variant.id()))
+        .set_default_schema_variant_id(ctx, Some(*schema_variant.id()))
         .await?;
 
     let mut secret_prop = Prop::new(ctx, "secret", PropKind::Integer).await?;
@@ -55,11 +55,12 @@ async fn docker_hub_credential(ctx: &DalContext) -> BuiltinsResult<()> {
         .pop()
         .ok_or(SchemaError::FuncNotFound(qual_func_name))?;
     let mut qual_prototype_context = QualificationPrototypeContext::new();
-    qual_prototype_context.set_schema_variant_id(*variant.id());
+    qual_prototype_context.set_schema_variant_id(*schema_variant.id());
 
     let _ = QualificationPrototype::new(ctx, *qual_func.id(), qual_prototype_context).await?;
 
-    let identity_func = BuiltinSchemaHelpers::setup_identity_func(ctx).await?;
+    let (identity_func_id, identity_func_binding_id, identity_func_binding_return_value_id) =
+        BuiltinSchemaHelpers::setup_identity_func(ctx).await?;
 
     let system_socket = Socket::new(
         ctx,
@@ -70,24 +71,24 @@ async fn docker_hub_credential(ctx: &DalContext) -> BuiltinsResult<()> {
         &DiagramKind::Configuration,
     )
     .await?;
-    variant.add_socket(ctx, system_socket.id()).await?;
+    schema_variant.add_socket(ctx, system_socket.id()).await?;
 
     let (_output_provider, mut output_socket) = ExternalProvider::new_with_socket(
         ctx,
         *schema.id(),
-        *variant.id(),
+        *schema_variant.id(),
         "docker_hub_credential",
         None,
-        identity_func.0,
-        identity_func.1,
-        identity_func.2,
+        identity_func_id,
+        identity_func_binding_id,
+        identity_func_binding_return_value_id,
         SocketArity::Many,
         DiagramKind::Configuration,
     )
     .await?;
     output_socket.set_color(ctx, Some(0x1e88d6)).await?;
 
-    variant.finalize(ctx).await?;
+    schema_variant.finalize(ctx).await?;
 
     // Note: I wasn't able to create a ui menu with two layers
     let diagram_kind = schema
@@ -109,15 +110,15 @@ async fn docker_image(ctx: &DalContext) -> BuiltinsResult<()> {
             None => return Ok(()),
         };
 
-    let (mut variant, root_prop) = SchemaVariant::new(ctx, *schema.id(), "v0").await?;
-    variant.set_color(ctx, Some(0xd61e8c)).await?;
+    let (mut schema_variant, root_prop) = SchemaVariant::new(ctx, *schema.id(), "v0").await?;
+    schema_variant.set_color(ctx, Some(0xd61e8c)).await?;
     schema
-        .set_default_schema_variant_id(ctx, Some(*variant.id()))
+        .set_default_schema_variant_id(ctx, Some(*schema_variant.id()))
         .await?;
     let mut attribute_context_builder = AttributeContext::builder();
     attribute_context_builder
         .set_schema_id(*schema.id())
-        .set_schema_variant_id(*variant.id());
+        .set_schema_variant_id(*schema_variant.id());
 
     let diagram_kind = schema
         .diagram_kind()
@@ -155,7 +156,7 @@ async fn docker_image(ctx: &DalContext) -> BuiltinsResult<()> {
         InternalProvider::new_explicit_with_socket(
             ctx,
             *schema.id(),
-            *variant.id(),
+            *schema_variant.id(),
             "docker_hub_credential",
             identity_func_id,
             identity_func_binding_id,
@@ -169,7 +170,7 @@ async fn docker_image(ctx: &DalContext) -> BuiltinsResult<()> {
     let (docker_image_external_provider, mut output_socket) = ExternalProvider::new_with_socket(
         ctx,
         *schema.id(),
-        *variant.id(),
+        *schema_variant.id(),
         "docker_image",
         None,
         identity_func_id,
@@ -190,7 +191,7 @@ async fn docker_image(ctx: &DalContext) -> BuiltinsResult<()> {
         &DiagramKind::Configuration,
     )
     .await?;
-    variant.add_socket(ctx, system_socket.id()).await?;
+    schema_variant.add_socket(ctx, system_socket.id()).await?;
 
     // Qualification Prototype
     let qual_func_name = "si:qualificationDockerImageNameInspect".to_string();
@@ -199,7 +200,7 @@ async fn docker_image(ctx: &DalContext) -> BuiltinsResult<()> {
         .pop()
         .ok_or(SchemaError::FuncNotFound(qual_func_name))?;
     let mut qual_prototype_context = QualificationPrototypeContext::new();
-    qual_prototype_context.set_schema_variant_id(*variant.id());
+    qual_prototype_context.set_schema_variant_id(*schema_variant.id());
 
     let _ = QualificationPrototype::new(ctx, *qual_func.id(), qual_prototype_context).await?;
 
@@ -212,7 +213,7 @@ async fn docker_image(ctx: &DalContext) -> BuiltinsResult<()> {
     let resource_sync_args = FuncBackendJsResourceSyncArgs::default();
     let resource_sync_args_json = serde_json::to_value(&resource_sync_args)?;
     let mut resource_sync_prototype_context = ResourcePrototypeContext::new();
-    resource_sync_prototype_context.set_schema_variant_id(*variant.id());
+    resource_sync_prototype_context.set_schema_variant_id(*schema_variant.id());
 
     let _prototype = ResourcePrototype::new(
         ctx,
@@ -222,11 +223,11 @@ async fn docker_image(ctx: &DalContext) -> BuiltinsResult<()> {
     )
     .await?;
 
-    variant.finalize(ctx).await?;
+    schema_variant.finalize(ctx).await?;
 
     let base_attribute_read_context = AttributeReadContext {
         schema_id: Some(*schema.id()),
-        schema_variant_id: Some(*variant.id()),
+        schema_variant_id: Some(*schema_variant.id()),
         ..AttributeReadContext::default()
     };
 
@@ -290,7 +291,7 @@ async fn docker_image(ctx: &DalContext) -> BuiltinsResult<()> {
     let title = "Docker Image Resource Refresh";
     let context = WorkflowPrototypeContext {
         schema_id: *schema.id(),
-        schema_variant_id: *variant.id(),
+        schema_variant_id: *schema_variant.id(),
         ..Default::default()
     };
     WorkflowPrototype::new(ctx, *func.id(), serde_json::Value::Null, context, title).await?;
