+++ lib/dal/src/builtins/schema/kubernetes/kubernetes_selector.rs
@@ -7,56 +7,38 @@ pub async fn create_selector_prop(
     ctx: &DalContext,
     parent_prop_id: PropId,
 ) -> BuiltinsResult<Prop> {
-    let mut selector_prop = BuiltinSchemaHelpers::create_prop(
+    let selector_prop = BuiltinSchemaHelpers::create_prop(
         ctx,
         "selector",
         PropKind::Object,
         Some(parent_prop_id),
-        None,
+        Some(doc_url(
+            "reference/kubernetes-api/common-definitions/label-selector/#LabelSelector",
+        )),
     )
     .await?;
-    selector_prop
-        .set_doc_link(
+
+    {
+        let match_labels_prop = BuiltinSchemaHelpers::create_prop(
             ctx,
+            "matchLabels",
+            PropKind::Map, // How to specify it as an array of strings?
+            Some(*selector_prop.id()),
             Some(doc_url(
                 "reference/kubernetes-api/common-definitions/label-selector/#LabelSelector",
             )),
         )
         .await?;
-
-    {
-        let mut match_labels_prop = BuiltinSchemaHelpers::create_prop(
-            ctx,
-            "matchLabels",
-            PropKind::Map, // How to specify it as an array of strings?
-            Some(*selector_prop.id()),
-            None,
-        )
-        .await?;
-        match_labels_prop
-            .set_doc_link(
-                ctx,
-                Some(doc_url(
-                    "reference/kubernetes-api/common-definitions/label-selector/#LabelSelector",
-                )),
-            )
-            .await?;
-        let mut match_labels_value_prop = BuiltinSchemaHelpers::create_prop(
+        let _match_labels_value_prop = BuiltinSchemaHelpers::create_prop(
             ctx,
             "labelValue",
             PropKind::String,
             Some(*match_labels_prop.id()),
-            None,
+            Some(doc_url(
+                "reference/kubernetes-api/common-definitions/label-selector/#LabelSelector",
+            )),
         )
         .await?;
-        match_labels_value_prop
-            .set_doc_link(
-                ctx,
-                Some(doc_url(
-                    "reference/kubernetes-api/common-definitions/label-selector/#LabelSelector",
-                )),
-            )
-            .await?;
     }
 
     Ok(selector_prop)
