+++ lib/dal/src/queries/ui_menus_list_for_diagram_kind.sql
