+++ lib/sdf/src/server/service/dev/save_builtin_func.rs
@@ -52,7 +52,7 @@ pub async fn save_builtin_func(
     func.set_code_plaintext(&ctx, request.code.as_deref())
         .await?;
 
-    dal::builtins::func::persist(&func).await?;
+    dal::builtins::func_persist(&func).await?;
 
     // Update the ctx with the account details for proper WS signaling
     ctx.update_from_request_context(request_ctx.build(request.visibility));
